requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.2
opencv-python==4.9.0.80
Pillow==10.2.0
tensorflow==2.16.1   # If using TensorFlow
torch==2.3.0         # If using PyTorch
torchvision==0.18.0  # If using PyTorch vision
scikit-learn==1.5.0
matplotlib==3.9.0
seaborn==0.13.2
